Create life_in_the_shell.py
This is the first implementation of Conway Game of Life that I was comfortable coding along. As stated in the first comment it is heavily inspired by Al Sweigart's one in his most famous book Automate the Boring Stuff.

This code serves as the base from which I will try to build more sophisticated visualizations and implementation, so take it as an exercise.